services/llm/pyproject.toml
[project]
name = "ekho-llm"
version = "0.1.0"
description = "LLM service for EKHO - Contextual translation with Gemini"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "ekho-core>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

[tool.setuptools.packages.find]
where = ["."]
include = ["llm*"]
